// autogpt_platform/frontend/src/app/api/__generated__/models/aPIKeyStatus.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export type APIKeyStatus = (typeof APIKeyStatus)[keyof typeof APIKeyStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const APIKeyStatus = {
  ACTIVE: "ACTIVE",
  REVOKED: "REVOKED",
  SUSPENDED: "SUSPENDED",
} as const;